# client/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : ""
  },
  "basePath" : "/",
  "schemes" : [ "http" ],
  "paths" : {
...
    "/typeclasses" : {
      "get" : {
        "tags" : [ "typeclasses" ],
        "summary" : "List",
        "description" : "",
        "operationId" : "list",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Typedescription"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Create",
        "description" : "",
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Typedescription"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Typedescription"
            }
          }
        }
      }
    },
    "/typeclasses/{id}" : {
      "get" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Read",
        "description" : "",
        "operationId" : "read",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Typedescription"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Update",
        "description" : "",
        "operationId" : "update",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Typedescription"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Typedescription"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Delete",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ]
      }
    },
    "/typeclasses/{typeclassid}/typevalues" : {
      "get" : {
        "tags" : [ "typeclasses" ],
        "summary" : "List global types",
        "description" : "",
        "operationId" : "getGlobalTypeForTypeClass",
        "parameters" : [ {
          "name" : "typeclassid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ]
      },
      "post" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Create global type",
        "description" : "",
        "operationId" : "createGlobalType",
        "parameters" : [ {
          "name" : "typeclassid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Globaltype"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Globaltype"
            }
          }
        }
      }
    },
    "/typeclasses/{typeclassid}/typevalues/{typevalueid}" : {
      "get" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Read global type",
        "description" : "",
        "operationId" : "getGlobalType",
        "parameters" : [ {
          "name" : "typeclassid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "typevalueid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Globaltype"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Update global type",
        "description" : "",
        "operationId" : "updateGlobalType",
        "parameters" : [ {
          "name" : "typeclassid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "typevalueid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Globaltype"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Globaltype"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "typeclasses" ],
        "summary" : "Delete global type",
        "description" : "",
        "operationId" : "deleteGlobalType",
        "parameters" : [ {
          "name" : "typeclassid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "typevalueid",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ]
      }
    }
  },
  "definitions" : {
...
    "Globaltype" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "typeid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "shortdescription" : {
          "type" : "string"
        },
        "longdescription" : {
          "type" : "string"
        },
        "typecode" : {
          "type" : "string"
        },
        "created" : {
          "$ref" : "#/definitions/Timestamp"
        },
        "lastmodified" : {
          "$ref" : "#/definitions/Timestamp"
        },
        "createdby" : {
          "type" : "integer",
          "format" : "int32"
        },
        "lastmodifiedby" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Timestamp" : {
      "properties" : {
        "nanos" : {
          "type" : "integer",
          "format" : "int32"
        },
        "time" : {
          "type" : "integer",
          "format" : "int64"
        },
        "year" : {
          "type" : "integer",
          "format" : "int32"
        },
        "month" : {
          "type" : "integer",
          "format" : "int32"
        },
        "date" : {
          "type" : "integer",
          "format" : "int32"
        },
        "hours" : {
          "type" : "integer",
          "format" : "int32"
        },
        "minutes" : {
          "type" : "integer",
          "format" : "int32"
        },
        "seconds" : {
          "type" : "integer",
          "format" : "int32"
        },
        "day" : {
          "type" : "integer",
          "format" : "int32"
        },
        "timezoneOffset" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "Typedescription" : {
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "howused" : {
          "type" : "string"
        }
      }
    },
...
  }
}